requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
openai==1.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
dj-database-url==2.1.0

